requirements.txt
requests==2.31.0
python-dotenv==1.0.0
web3==6.11.1
pymongo==4.6.1
python-dateutil==2.8.2 